Rebasing all modular content tests into a single test file
Modularity is a wide reaching feature within our repo. It is an
evolving topic with a lot of feature regression and updates.

Moving all of the tests into a single file to enable ease of
finding relevant tests.

The topics currently included with modular content testing are:
* Upload
* Sync
* Defaults
* Copy
* Errata
* Applicability
* Errata Copy

Run time is ~6mins.

No tests are marked at this time for pytest marking runs.
